apps/members/views.py
"""
Views for members app.
"""
from rest_framework import generics, status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny, IsAuthenticated
from rest_framework.response import Response
from django.contrib.auth.models import User
from rest_framework_simplejwt.tokens import RefreshToken

class RegisterView(generics.CreateAPIView):
    """
    User registration view.
    """
    queryset = User.objects.all()
    permission_classes = [AllowAny]
    
    def create(self, request, *args, **kwargs):
        # Basic registration - to be expanded
        return Response({'message': 'Registration endpoint - to be implemented'})

class ProfileView(generics.RetrieveAPIView):
    """
    Get user profile.
    """
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        return Response({'message': 'Profile endpoint - to be implemented'})

class ProfileUpdateView(generics.UpdateAPIView):
    """
    Update user profile.
    """
    permission_classes = [IsAuthenticated]
    
    def put(self, request):
        return Response({'message': 'Profile update endpoint - to be implemented'})

class LogoutView(generics.GenericAPIView):
    """
    Logout view.
    """
    permission_classes = [IsAuthenticated]
    
    def post(self, request):
        return Response({'message': 'Logged out successfully'})

class MemberDirectoryView(generics.ListAPIView):
    """
    Member directory view.
    """
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        return Response({'message': 'Member directory - to be implemented'})
