requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# File handling & Media
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# Payment processing
stripe==7.8.0
requests==2.31.0

# Email & Communication
django-ses==3.5.0
celery==5.3.4
redis==5.0.1

# API Documentation
drf-spectacular==0.26.5

# Development & Testing
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.6.0
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Utilities
python-decouple==3.8
django-filter==23.4
django-taggit==4.0.0
django-ckeditor==6.7.0
